+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 # Unreleased
 
+No current changes
+
+# v4.10.3
+
 - Option to hide interfaces from get
 - Simplified `**kwargs` arguments in CLI
 - Adds save/load_attribute helper